refactor: split Next use cache runtime bridge
